refactor(photo-annotator): remove dead SVG infrastructure and tool dispatcher
Delete all tool*.ts files, geometry.test.ts, render.test.ts, and render.ts.
Moved ANNOTATION_FONT_FAMILY and helper functions to canvasRenderer.ts.

Files deleted:
- tools/ directory (20 files: 10 tool implementations + 10 test files)
- geometry.test.ts
- render.test.ts
- render.ts

Files updated:
- canvasRenderer.ts: Added calculateCalloutEffectiveFontSize, wrapTextForCanvas, ANNOTATION_FONT_FAMILY
- PhotoAnnotator.tsx: Import ANNOTATION_FONT_FAMILY from canvasRenderer instead of render

geometry.ts kept: clamp, nearestBoxEdgePoint, distance still used.
